Removed payment service, added action service
