+++ helm/ffc-ce-admin-web/jenkins-aws.yaml
@@ -5,7 +5,6 @@ service:
 container:
   allowPrivilegeEscalation: false
   imagePullPolicy: Always
-  paymentUrl: http://ffc-ce-payment-orchestrator.ffc-ce
 ingress:
   endPoint: ffc-ce-admin-web
   server: ffc.aws-int.defra.cloud
