+++ server/services/actionService.js
@@ -1,0 +1,6 @@
+
+async function performAction (email) {
+  return { success: true }
+}
+
+module.exports = { performAction }
