--- server/views/actions.njk
@@ -1,7 +1,0 @@
-{% extends 'layout.njk' %}
-
-{% block content %}
-  <div class="govuk-body">
-    Select actions you wish to enable / disable
-  </div>
-{% endblock %}
