+++ server/views/home.njk
@@ -29,11 +29,11 @@
       }) }}
 
 
-    <form method="POST" action="/actions">
+    <form method="POST" action="/action">
       {{ govukButton({
         classes: "govuk-!-margin-top-2 govuk-!-margin-bottom-8",
         isStartButton: true,
-        text: "Configure actions"
+        text: "Sample Action 1"
       }) }}
     </form>
       <h2 class="govuk-heading-m">Before you start</h2>
