+++ test/services/actionService.test.js
@@ -1,0 +1,10 @@
+const actionService = require('../../server/services/actionService')
+
+describe('actionService', () => {
+  test('action succeeds', () => {
+    const result = actionService.performAction()
+    expect(result).toEqual(
+      expect.objectContains({ success: true })
+    )
+  })
+})
